Fix type of Message.key to be optional string
key is not a number, it's a string. See KeyedMessage constructor, above.

(However, if `encoding` is specified as "buffer" in ConsumerOptions then `key` and `value` are actually `Buffer` rather than `string`. I'm not sure if it's possible to specify this better than `string|Buffer` since it may not be knowable at compile time.)